+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 * Possible StackOverflow crash regarding KO and Stop.
 * Aero damage being ridiculously high.
 * Keyblade damage dealt to other players.
+* Magic Reaction Commands not working.